+++ envelope/deferred_jobs/tests/test_async_signals.py
@@ -13,6 +13,7 @@ from rq import SimpleWorker
 from envelope import WS_INCOMING
 from envelope.channels.messages import Subscribe
 from envelope.channels.messages import Subscribed
+from envelope.deferred_jobs.async_signals import queue_deferred_job
 from envelope.logging import getEventLogger
 from envelope.messages.ping import Ping
 from envelope.models import Connection
@@ -135,9 +136,7 @@ class DeferredJobsAsyncSignalsTests(TestCase):
         self.assertIsInstance(connection.last_action, datetime)
 
     async def test_queue_deferred_job(self):
-
         self.mock_consumer.user_pk = self.user.pk
-
         msg = Subscribe(
             mm={"user_pk": self.user.pk, "env": WS_INCOMING},
             channel_type="user",
@@ -148,17 +147,32 @@ class DeferredJobsAsyncSignalsTests(TestCase):
             "django_rq.queues.get_redis_connection",
             return_value=self.fake_redis_conn,
         ):
-            from envelope.deferred_jobs.async_signals import queue_deferred_job
-
             await queue_deferred_job(consumer=self.mock_consumer, message=msg)
 
         queue = self.fake_redis_queue(name="default")
         worker = SimpleWorker([queue], connection=self.fake_redis_conn)
         result = await sync_to_async(worker.work)(burst=True)
         self.assertTrue(result)
+        self.assertTrue(msg.post_queue.called)
+        self.assertIn("job", msg.post_queue.mock_calls[0].kwargs)
+        self.assertIn("consumer", msg.post_queue.mock_calls[0].kwargs)
+
+    async def test_post_queue_sends_subscribe_queued(self):
+        self.mock_consumer.user_pk = self.user.pk
+        msg = Subscribe(
+            mm={"user_pk": self.user.pk, "env": WS_INCOMING},
+            channel_type="user",
+            pk=self.user.pk,
+        )
+        with patch(
+            "django_rq.queues.get_redis_connection",
+            return_value=self.fake_redis_conn,
+        ):
+            await queue_deferred_job(consumer=self.mock_consumer, message=msg)
         self.assertTrue(self.mock_consumer.ws_out)
         response = self.mock_consumer.ws_out[0]
         self.assertIsInstance(response, Subscribed)
+        self.assertEqual(response.QUEUED, response.mm.state)
         self.assertEqual(
             {
                 "app_state": None,
@@ -168,9 +182,6 @@ class DeferredJobsAsyncSignalsTests(TestCase):
             },
             response.data.dict(),
         )
-        self.assertTrue(msg.post_queue.called)
-        self.assertIn("job", msg.post_queue.mock_calls[0].kwargs)
-        self.assertIn("consumer", msg.post_queue.mock_calls[0].kwargs)
 
     async def test_queue_deferred_job_wrong_msg_type(self):
         self.mock_consumer.user_pk = self.user.pk
@@ -179,8 +190,6 @@ class DeferredJobsAsyncSignalsTests(TestCase):
             "django_rq.queues.get_redis_connection",
             return_value=self.fake_redis_conn,
         ):
-            from envelope.deferred_jobs.async_signals import queue_deferred_job
-
             await queue_deferred_job(consumer=self.mock_consumer, message=Ping())
 
         queue = self.fake_redis_queue(name="default")